// libs/common/src/utils/books/enum/category.enum.ts
export enum CategoryUtilsEnum {
  ALL = 'All',
  ACTION = 'Action',
  ADVENTURE = 'Adventure',
  BUSINESS = 'Business',
  COMEDY = 'Comedy',
  CRAFTS = 'Crafts',
  CRIME = 'Crime',
  DRAMA = 'Drama',
  GUIDE = 'Guide',
  HEALING = 'Healing',
  HUMOR = 'Humor',
  JOURNAL = 'Journal',
  MUSIC = 'Music',
  ROMANTIC = 'Romantic',
  SPORTS = 'Sports',
  TRAVEL = 'Travel',
}

export enum ESortBooksQuery {
  QUANTITY_ASC = 'quantity_asc',
  QUANTITY_DESC = 'quantity_desc',
  PRICE_ASC = 'price_asc',
  PRICE_DESC = 'price_desc',
}